MINOR: http-fetch: Add pathq sample fetch
The pathq sample fetch extract the relative URI of a request, i.e the path with
the query-string, excluding the scheme and the authority, if any. It is pretty
handy to always get a relative URI independently on the HTTP version. Indeed,
while relative URIs are common in HTTP/1.1, in HTTP/2, most of time clients use
absolute URIs.

This patch may be backported to 2.2.

(cherry picked from commit e720c32b78566ce6f5f610f6e14b09b5b774610d)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>