MINOR: server: Improve log message sent when server address is updated
When the server address is set for the first time, the log message is a bit ugly
because there is no old ip address to report. Thus in the log, we can see :

  PX/SRV changed its IP from  to A.B.C.D by DNS additional record.

Now, when this happens, "(none)" is reported :

  PX/SRV changed its IP from (none) to A.B.C.D by DNS additional record.

This patch may be backported to 2.2.

(cherry picked from commit b0b7607a5482cb7092e82e51cdbc17ad6efdc2f3)
Signed-off-by: Willy Tarreau <w@1wt.eu>