BUG/MINOR: ssl/crt-list: crt-list could end without a \n
Since the refactoring of the crt-list, the same function is used to
parse a crt-list file and a crt-list line on the CLI.

The assumption that a line on the CLI and a line in a file is finished
by a \n was made. However that is potentialy not the case with a file
which does not finish by a \n.

This patch fixes issue #860 and must be backported in 2.2.

(cherry picked from commit 0354b658f061d00d5ab4b728d7deeff2c8f1503a)
Signed-off-by: Willy Tarreau <w@1wt.eu>