BUG/MEDIUM: listeners: do not pause foreign listeners
There's a nasty case with listeners that belong to foreign processes.
If a proxy is defined this way:

     global
         nbproc 2

     frontend f
         bind :1111 process 1
         bind :2222 process 2

and if stats expose-fd listeners is set, the listeners' FDs will not
be closed on the processes that don't use them. At this point it's not
a big deal, except that they're shared between processes and that a
"disable frontend f" issued on one process will pause all of them and
cause the other process to see accept() fail, turning its own listener
to state LI_LIMITED to try to leave it some time to recover. But it
will never recover, even after an enable.

The root cause of the issue is that the ZOMBIE state doesn't cover
this situation since it's only for a proxy being entirely bound to a
process.

What we do here to address this is that we refrain from pausing a
file descriptor that belongs to a foreign process in pause_listener().
This definitely solves the problem. A similar test is present in
resume_listener() and is the reason why the FD doesn't recover upon the
"enable" action by the way.

This ought to be backported to 1.8 where seamless reload was integrated.

The config above should be sufficient to validate that the fix works;
after a pair of "disable/enable frontend" no process will handle the
traffic to one of the ports anymore.

(cherry picked from commit 02e1975c29eb19da98205164f23154f903802644)
[wt: context adjustment; bind_proc is in l->bind_conf in 2.2; test ok]
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit c7239d5da5801f2238b56601bfae29441153d4b8)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit efb5220d83985aa93ac9174d3e536c1d7999021d)
Signed-off-by: Willy Tarreau <w@1wt.eu>