DOC: ssl: crt-list negative filters are only a hint
This issue was reported in #818 but can't be fixed in this version
because there is too much dependencies to do it.

This patch explains the real behavior of the crt-list negative filters
and how to use them correctly.

This is specific to versions prior to 2.1 and must be backported in
all older branches.

(cherry picked from commit b0b703faa7a38f7f2b0f418a02cc83e044962bdb)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>